+++ uni-stark/src/lib.rs
@@ -4,8 +4,10 @@
 
 extern crate alloc;
 
+mod check_constraints;
 mod config;
 mod folder;
+mod preprocessed;
 mod proof;
 mod prover;
 mod symbolic_builder;
@@ -13,11 +15,10 @@ mod symbolic_expression;
 mod symbolic_variable;
 mod verifier;
 
-mod check_constraints;
-
 pub use check_constraints::*;
 pub use config::*;
 pub use folder::*;
+pub use preprocessed::*;
 pub use proof::*;
 pub use prover::*;
 pub use symbolic_builder::*;
