+++ uni-stark/tests/mul_fib_pair.rs
@@ -13,8 +13,7 @@ use p3_matrix::dense::RowMajorMatrix;
 use p3_merkle_tree::MerkleTreeMmcs;
 use p3_symmetric::{PaddingFreeSponge, TruncatedPermutation};
 use p3_uni_stark::{
-    PreprocessedVerifierKey, StarkConfig, prove_with_preprocessed, setup_preprocessed,
-    verify_with_preprocessed,
+    StarkConfig, prove_with_preprocessed, setup_preprocessed, verify_with_preprocessed,
 };
 use rand::SeedableRng;
 use rand::rngs::SmallRng;
@@ -203,19 +202,12 @@ fn test_mul_fib_pair() {
 
     let air = MulFibPAir::new(num_rows);
     let degree_bits = 10; // log2(1024)
-    let preprocessed = setup_preprocessed::<MyConfig, _>(&config, &air, degree_bits);
+    let (preprocessed_prover_data, preprocessed_vk) =
+        setup_preprocessed::<MyConfig, _>(&config, &air, degree_bits).unwrap();
 
-    let proof = prove_with_preprocessed(&config, &air, trace, &[], preprocessed.as_ref());
+    let proof = prove_with_preprocessed(&config, &air, trace, &[], Some(&preprocessed_prover_data));
 
-    let preprocessed_vk = preprocessed
-        .as_ref()
-        .map(|pp| PreprocessedVerifierKey::<MyConfig> {
-            width: pp.width,
-            degree_bits: pp.degree_bits,
-            commitment: pp.commitment,
-        });
-
-    verify_with_preprocessed(&config, &air, &proof, &[], preprocessed_vk.as_ref())
+    verify_with_preprocessed(&config, &air, &proof, &[], Some(&preprocessed_vk))
         .expect("verification failed");
 }
 
@@ -228,28 +220,22 @@ fn test_tampered_preprocessed_fails() {
     let degree_bits = 10; // log2(1024)
 
     // Prover uses the correct AIR for preprocessed setup.
-    let preprocessed = setup_preprocessed::<MyConfig, _>(&config, &air, degree_bits);
-    let proof = prove_with_preprocessed(&config, &air, trace, &[], preprocessed.as_ref());
+    let (preprocessed_prover_data, _) =
+        setup_preprocessed::<MyConfig, _>(&config, &air, degree_bits).unwrap();
+    let proof = prove_with_preprocessed(&config, &air, trace, &[], Some(&preprocessed_prover_data));
 
     // Verifier uses a *tampered* AIR to derive the preprocessed commitment, which should
     // not match the one used in the proof.
     let tampered_air = MulFibPAir::with_tampered_preprocessed(num_rows, 3);
-    let tampered_preprocessed =
-        setup_preprocessed::<MyConfig, _>(&config, &tampered_air, degree_bits);
-    let preprocessed_vk = tampered_preprocessed
-        .as_ref()
-        .map(|pp| PreprocessedVerifierKey::<MyConfig> {
-            width: pp.width,
-            degree_bits: pp.degree_bits,
-            commitment: pp.commitment,
-        });
+    let (_, tampered_preprocessed_vk) =
+        setup_preprocessed::<MyConfig, _>(&config, &tampered_air, degree_bits).unwrap();
 
     let result = verify_with_preprocessed(
         &config,
         &tampered_air,
         &proof,
         &[],
-        preprocessed_vk.as_ref(),
+        Some(&tampered_preprocessed_vk),
     );
 
     assert!(
